[tests][mmp] Move TypeDescriptor (regression) tests to unit tests
Also move some tests to share more (now and in future commits)
with iOS linker-related tests.

Part of https://github.com/xamarin/xamarin-macios/issues/4975
