+++ tests/linker/mac/link all/link all-mac.csproj
@@ -78,6 +78,9 @@
     <Compile Include="..\..\..\common\TestRuntime.cs">
       <Link>TestRuntime.cs</Link>
     </Compile>
+    <Compile Include="..\..\CommonLinkAllTest.cs">
+      <Link>CommonLinkAllTest.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\external\guiunit\src\framework\GuiUnit_xammac_mobile.csproj">
